chore(plot): rename callback_scheduling_source to callback_scheduling
